Move model bundles into Documents/models
The new behaviour is that the application checks when it is loaded to
see if a Documents/models directory already exists. If not, it copies
over the initial models/ directory in which the build-time bundles are
present.

This sets up for future work in which we will let users download model
bundles directly onto their devices (without requiring them to be
present at build time).